[Report] Refactors report
I changed the way things are displayed to have more granular
information on what's Serverless, what's Stack, and where they're
coming from. We now have a list of Stack and Serverless APIs and which
ones are being tested. I also added the difference between the APIs in
the Elasticsearch JSON spec that are not in the
elasticsearch-specification.