Refactors reporter.
As the script grew and became more complex, it needed some
refactoring. I updated it to hopefully become more clear and easy to
maintain.

The reporter has less responsibilty and each endpoint becomes an object
which deals with the responsibility of setting up the availability and
tested values. I fixed some issues on the way, such as the calculation
for coverage in Stack which was reporting a lower percentage (since it
was being compared to the JSON spec instead of
elasticsearch-specification).

I also improved the reading of files, so now the script runs in a
couple of seconds instead of a few seconds like before (it was not
optimized at all).