Adds back ml.flush_job test
I tested this both in Stack and Serverless and it's working ok.
The first time running on Serverless it took some time, so maybe
it has to do with what I've done for Inference tests in #83.

We can test it for other clients and if it's still an issue we can
come back to it.

Resolves #31